# Use sync_file_range to optimize fsync if possible