--- a/src/rio.c
+++ b/src/rio.c
@@ -46,7 +46,7 @@
 
 
 #include "fmacros.h"
-#include <assert.h>
+#include "redisassert.h"
 #include <string.h>
 #include <stdio.h>
 #include <unistd.h>
@@ -112,6 +112,8 @@ static size_t rioFileWrite(rio *r, const void *buf, size_t len) {
     if (!r->io.file.autosync) return fwrite(buf,len,1,r->io.file.fp);
 
     size_t nwritten = 0;
+    /* Incrementally write data to the file, avoid a single write larger than the autosync
+       threshold (so that the kernel's buffer cache never has too many dirty pages at once). */
     while (len != nwritten) {
         assert(r->io.file.autosync > r->io.file.buffered);
         size_t nalign = (size_t)(r->io.file.autosync - r->io.file.buffered);