--- a/src/rio.c
+++ b/src/rio.c
@@ -111,8 +111,9 @@ static size_t rioFileWrite(rio *r, const void *buf, size_t len) {
     if (!r->io.file.autosync) return fwrite(buf,len,1,r->io.file.fp);
 
     size_t nwritten = 0;
-    /* Incrementally write data to the file, avoid a single write larger than the autosync
-       threshold (so that the kernel's buffer cache never has too many dirty pages at once). */
+    /* Incrementally write data to the file, avoid a single write larger than
+     * the autosync threshold (so that the kernel's buffer cache never has too
+     * many dirty pages at once). */
     while (len != nwritten) {
         serverAssert(r->io.file.autosync > r->io.file.buffered);
         size_t nalign = (size_t)(r->io.file.autosync - r->io.file.buffered);