@oranagra 2021-08-25
@yoav-steinberg please take a look too.
in RL we have a similar change which we never got to push a PR for (i think) that does this check to decide if fflush/fsync should be called (also split big writes to a loop of small ones):
```c
        if (r->io.file.autosync && (r->processed_bytes + bytes_written)/r->io.file.autosync < (r->processed_bytes + bytes_written + bytes_to_write)/r->io.file.autosync)
```