@ShooterIT 2021-08-30
I forget update comments about `rdb-save/aof-rewrite-incremental-fsync` in redis.conf