@oranagra 2021-08-30
@ShooterIT this PR is ready to be merged now, right?
please have a look at my edit of the top comment, to see if it's ok and can be used as squash-merge commit comment.

do you think this should be mentioned in the release notes? it doesn't introduce any actual change, and on modern storage the the improvement is certainly welcome, but not a game changer.