@ShooterIT 2021-08-30
Thanks @oranagra  the top comment looks good to me.
For release notes, I am not sure, it is decided by you. I prefer to mention it since we let  serializing key/values and syncing file parallelize, so there is a performance promotion, of course, on modern storage, this change doesn't bring much promotion.